Players can register using game code
Game code is a prefix of the gameId (the entire Game ID works too).

Need to test this works on S3, works in local persistence.

Some brute force issues here, although games are only registerable
briefly so perhaps not too big a deal. Requires at least 4 chars. This
is a balancing act between usabililty and security.