Refactor game state to include player points
PlayerSummary is used to hold basic player info and includes the
points a player has been awarded.

For now the initial game state is created with empty points and things
compile, next up is to add the behaviour that keeps the player points
in the game state in sync with each player's persisted `PlayerState`.