Implement shroud for "loading backend"
Design not yet A+, but it works ok.

Real page exists behind the message as a blueprint for other loading
lifecycle pages. This allows e.g. form data to persist behind the
loading screen to make error handling of a submission neater.

The "backend loading" message is likely to go away completely at some
point, just a PoC.