Start proper lifecycle for creation
TODO: Simplify create lifecycle into a message for the screen, and a
msg for handling the API response.

TODO: Move error state into the lifecycle so it doesn't persist
forever and break everything.

TODO: Error styles.

TODO: Refactor shroud so it's always there with opacity: 0 and
display: none, add transition to opacity so it fades in.