Introduce game rounds
The game state keeps track of the game round, which is the buyer (with
role) and all the pitches that have been made in this round.

We swap the simple `buyer` field for a more complete `round`
datastructure in the game state and add round information to the data
sent to clients.