sbt upgrade broke the integration tests?!?!
Reading from resources seems to have changed in the later version of
sbt. but it isn't trivial to work out why. The tests fail
intermittently with the new version of sbt, which is super annoying.

Rather than continue to waste time solving it, I've rolled back to the
version of sbt that definitely works.