Add lobbyPing endpoint to API
This is the same as ping, except it collects all the registered
players that haven't yet been added to the game state. This operation
involves more IO than the standard ping so it's only available to the
game creator for now. This is probably fine from UX point of view,
sice it's only really them that need to know this information.

Next up, adding it to the UI.

Note: The resources changes came from the sbt faffing (see previous
commit). This approach is probably a bit better than what was there
before, since it now certainly won't duplicate the effort of loading
the resources. It's a minor point though for such small files.