Add lobby players to creator waiting screen
This wires in the lobbyPing API endpoint so that players that have
joined a new game are visible to the creator before the game
starts. The CreatorWaiting screen is slightly simplified replacing the
old player add icon with the list of joined players.

Also fixes a couple of lifecyle bugs to do with leaving games that
came up during testing of this feature.