Groundwork for allowing buyer to choose from multiple roles
The data-model has been extended to include a buyer choice field, that
may include roles for the new buyer to choose from. It also includes
stub functions in the API to implement this functionality in a
backwards-compatible way. The existing 'become buyer' functionality
can be preserved for now, with new functionality to provide the
multi-step buyer flow that includes a role choice step.