Add URL routing / navigation to support joining games from links
This is part of enabling support for external services that set up
games (e.g. 🚀🦀). A correct link will take the player straight to the
join screen and this works for player joins as well as host joins.

Quite a few changes were required to support this change, including
switching from Browser.Element to Browser.Application.

Some fiddly logic has tests, so the frontend test step is added to the
build definition as well.

This is not quite ready for release, a bit more polish and testing
would be sensible.