Add screenshot diffs to WATs
These diffs are generated by comparing the current run's screenshots
to a set of reference screenshots that are checked into version
control.

There's some noise on the screenshots since the game randomly chooses
cards, roles and the game's code. This means the diffs can't be
automated, but they're still useful for a visual inspection prior to
release.