Switch to Cats Effect
Drops both the custom `EitherT[Future, L, R]` wrapper and the use of
Future, replace them with Cats' EitherT Cats Effect's IO respectively.

We preserve `Attempt[A]` as a convenient type alias and for some
consistency, but this still touches a large amount of the codebase.

Tested locally, additional testing will be required in the live AWS
environment. It'd also be good to check if there are any meaningful
differences in the application's (production) performance.